Define the system message for json output in the SystemMessage class
