+++ src/services/SystemMessage.js
@@ -1,0 +1,22 @@
+export default class SystemMessage {
+
+  static systemMessages() {
+    return [{ role: "system", content: `Your response should be entirely valid json with no other content outside of the json.
+Your reponse should be able to be parsed as json.
+Respond with valid json only. 
+Do not include file contents or any other words before or after the json.
+Do not respond with anything but json.
+The json should be an object with an "operations" key. 
+The "operations" key should be array of objects. 
+Each object should represent a file that should be created, updated, or deleted. 
+Each object should have three keys: "crudOperation", "filePath", and "fileContents".
+The "crudOperation" value should contain the operation that you would like to perform for the given file. The "crudOperation" value should be "create", "update", or "delete".
+The "filePath" value should contain the path to the file. 
+The "fileContents" value should be the contents of the file if the file is being created or updated - if the file is being deleted then the "fileContents" key can be omitted.
+Make sure that the "fileContents" value is delimitted correctly as a json string.
+Your json response must always be valid json.
+Only include changed files in your response. 
+Don't abbreviate file contents - include the whole file for the "fileContents" value.` }]
+  }
+  
+}
